airflow/requirements.txt
apache-airflow[celery]==2.6.1
psycopg2-binary==2.9.3
Faker==8.10.3
pandas==1.3.3
duckdb==1.1.3
celery==5.2.7
redis==3.5.3
setuptools==66.1.1
python-daemon==3.0.1
flower==1.2.0
airflow-dbt==0.4.0
